collector/requirements.txt
amqp==5.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
googletrans==3.1.0a0
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
rfc3986==1.5.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
# google-generativeai==0.8.3
alembic==1.14.1
google-cloud-texttospeech==2.26.0
langchain==0.2.5
langchain-core==0.2.8
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.1
langgraph==0.1.5
pydantic==1.10.22
pydantic_core==2.27.2
pytz==2025.2
gTTS==2.5.4